# schemas/openapi.json
{
  "components": {
    "schemas": {
      "CsvExportRequest": {
        "additionalProperties": false,
        "properties": {
          "all_scalar_curves": {
            "default": false,
            "title": "All Scalar Curves",
            "type": "boolean"
          },
          "curve_ids": {
            "items": {
              "type": "string"
            },
            "title": "Curve Ids",
            "type": "array"
          },
          "destination_path": {
            "minLength": 1,
            "title": "Destination Path",
            "type": "string"
          }
        },
        "required": [
          "destination_path"
        ],
        "title": "CsvExportRequest",
        "type": "object"
      },
...
      "CursorValueRequest": {
        "additionalProperties": false,
        "properties": {
          "curve_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Curve Ids",
            "type": "array"
          },
          "index": {
            "title": "Index",
            "type": "number"
          }
        },
        "required": [
          "curve_ids",
          "index"
        ],
        "title": "CursorValueRequest",
        "type": "object"
      },
      "CursorValueResponse": {
        "additionalProperties": false,
        "properties": {
          "requested_index": {
            "title": "Requested Index",
            "type": "number"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/CursorCurveValue"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "requested_index",
          "values"
        ],
        "title": "CursorValueResponse",
        "type": "object"
      },
...
      "DatasetViewSettings": {
        "additionalProperties": false,
        "properties": {
          "manual_anchor_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Anchor Index"
          },
          "manual_anchor_timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Anchor Timestamp"
          },
          "time_display_mode": {
            "$ref": "#/components/schemas/TimeDisplayMode",
            "default": "elapsed"
          },
          "time_zone": {
            "$ref": "#/components/schemas/TimeZoneMode",
            "default": "utc"
          }
        },
        "title": "DatasetViewSettings",
        "type": "object"
      },
      "DatasetViewSettingsUpdate": {
        "additionalProperties": false,
        "properties": {
          "manual_anchor_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Anchor Index"
          },
          "manual_anchor_timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Anchor Timestamp"
          },
          "time_display_mode": {
            "$ref": "#/components/schemas/TimeDisplayMode"
          },
          "time_zone": {
            "$ref": "#/components/schemas/TimeZoneMode"
          }
        },
        "required": [
          "time_display_mode",
          "time_zone"
        ],
        "title": "DatasetViewSettingsUpdate",
        "type": "object"
      },
...
      "DocumentSummary": {
        "additionalProperties": false,
        "properties": {
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DocumentDatasetSummary"
            },
            "title": "Datasets",
            "type": "array"
          },
          "field_name": {
            "title": "Field Name",
            "type": "string"
          },
          "file_size_bytes": {
            "title": "File Size Bytes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modified": {
            "title": "Modified",
            "type": "boolean"
          },
          "preserved_object_count": {
            "title": "Preserved Object Count",
            "type": "integer"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          },
          "scalar_curve_count": {
            "title": "Scalar Curve Count",
            "type": "integer"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "source_format": {
            "$ref": "#/components/schemas/SourceFormat"
          },
          "source_version": {
            "title": "Source Version",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "source_file",
          "source_format",
          "source_version",
          "field_name",
          "file_size_bytes",
          "scalar_curve_count",
          "saved",
          "modified",
          "datasets",
          "preserved_object_count",
          "warnings"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobAcceptedResponse",
        "type": "object"
      },
...
      "MetadataObjectPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MetadataObjectSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "items"
        ],
        "title": "MetadataObjectPage",
        "type": "object"
      },
...
      "OpenDocumentRequest": {
        "additionalProperties": false,
        "properties": {
          "index_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Candidate Id"
          },
          "max_preview_points": {
            "default": 800,
            "maximum": 2000.0,
            "minimum": 100.0,
            "title": "Max Preview Points",
            "type": "integer"
          },
          "source_path": {
            "minLength": 1,
            "title": "Source Path",
            "type": "string"
          }
        },
        "required": [
          "source_path"
        ],
        "title": "OpenDocumentRequest",
        "type": "object"
      },
      "PackageVerificationResponse": {
        "additionalProperties": false,
        "properties": {
          "asset_count": {
            "title": "Asset Count",
            "type": "integer"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "package_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Version"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "package_version",
          "asset_count",
          "errors"
        ],
        "title": "PackageVerificationResponse",
        "type": "object"
      },
      "SaveDocumentRequest": {
        "additionalProperties": false,
        "properties": {
          "destination_path": {
            "minLength": 1,
            "title": "Destination Path",
            "type": "string"
          }
        },
        "required": [
          "destination_path"
        ],
        "title": "SaveDocumentRequest",
        "type": "object"
      },
      "ScalarPreviewPageRequest": {
        "additionalProperties": false,
        "properties": {
          "curve_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Curve Ids",
            "type": "array"
          },
          "page": {
            "default": 0,
            "minimum": 0.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 100,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          }
        },
        "title": "ScalarPreviewPageRequest",
        "type": "object"
      },
      "ScalarVisibleRangeRequest": {
        "additionalProperties": false,
        "properties": {
          "curve_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "minItems": 1,
            "title": "Curve Ids",
            "type": "array"
          },
          "index_maximum": {
            "title": "Index Maximum",
            "type": "number"
          },
          "index_minimum": {
            "title": "Index Minimum",
            "type": "number"
          },
          "point_budget": {
            "default": 8000,
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Point Budget",
            "type": "integer"
          },
          "viewport_height": {
            "maximum": 10000.0,
            "minimum": 100.0,
            "title": "Viewport Height",
            "type": "integer"
          }
        },
        "required": [
          "curve_ids",
          "index_minimum",
          "index_maximum",
          "viewport_height"
        ],
        "title": "ScalarVisibleRangeRequest",
        "type": "object"
      },
      "SourceFormat": {
        "enum": [
          "LAS",
          "DLIS",
          "WITSML"
        ],
        "title": "SourceFormat",
        "type": "string"
      },
...
      "TimeDisplayMode": {
        "enum": [
          "elapsed",
          "clock"
        ],
        "title": "TimeDisplayMode",
        "type": "string"
      },
...
      "TimeZoneMode": {
        "enum": [
          "utc",
          "local"
        ],
        "title": "TimeZoneMode",
        "type": "string"
      },
...
      "VerifyPackageRequest": {
        "additionalProperties": false,
        "properties": {
          "package_path": {
            "minLength": 1,
            "title": "Package Path",
            "type": "string"
          }
        },
        "required": [
          "package_path"
        ],
        "title": "VerifyPackageRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Well Log Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/documents/open": {
      "post": {
        "operationId": "openDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/verify": {
      "post": {
        "operationId": "verifyPackage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPackageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageVerificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Package",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "operationId": "getDocument",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/close": {
      "post": {
        "operationId": "closeDocument",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/datasets/{dataset_id}/cursor-values": {
      "post": {
        "operationId": "getCursorValues",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CursorValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorValueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cursor Values",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/datasets/{dataset_id}/export-csv": {
      "post": {
        "operationId": "exportDatasetCsv",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CsvExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Dataset Csv",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/datasets/{dataset_id}/scalar/preview": {
      "post": {
        "operationId": "getScalarPreviewPage",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScalarPreviewPageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scalar Preview Page",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/datasets/{dataset_id}/scalar/visible-range": {
      "post": {
        "operationId": "getScalarVisibleRange",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScalarVisibleRangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scalar Visible Range",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/datasets/{dataset_id}/view-settings": {
      "put": {
        "operationId": "updateDatasetViewSettings",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetViewSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetViewSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Dataset View Settings",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/metadata-objects": {
      "get": {
        "operationId": "listMetadataObjects",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataObjectPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Metadata Objects",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/metadata-objects/{object_id}": {
      "get": {
        "operationId": "getMetadataObject",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataObjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Metadata Object",
        "tags": [
          "document data"
        ]
      }
    },
    "/api/v1/documents/{document_id}/save-as": {
      "post": {
        "operationId": "saveDocumentAs",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Document As",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}